I am now able to communicate with the SPI Memory chip and retrieve the status byte. I verified that the status byte is being retrieved correctly by sending the WREN and WRDI commands to the chip and observing the results. I will need to test the remainder of the functionality soon.

+++ src/spimem.h
@@ -59,7 +59,7 @@ uint32_t spi_mem_buff_sect_num;	// Current sector number of what is loaded into 
 /*		Fuction Prototypes				*/
 void spimem_initialize(void);
 uint32_t spimem_write(uint8_t spi_chip, uint32_t addr, uint32_t* data_buff, uint32_t size);
-uint32_t spimem_read(uint32_t spi_chip, uint32_t addr, uint16_t* read_buff, uint32_t size);
+uint32_t spimem_read(uint32_t spi_chip, uint32_t addr, uint8_t* read_buff, uint32_t size);
 uint32_t check_page(uint32_t page_num);
 uint32_t check_if_wip(uint32_t spi_chip);
 uint32_t get_page(uint32_t addr);
